School-Web_programming
The project is the realization of a multi-page site on the topic of Friedl Beekeeping. The site contains a realized search, the gallery was realized via Ajax by reading data from an XML file, a survey based on which cookies are created, a description and explanation of certain terms related to the topic itself. This project required learning content management using procedural JavaScript and the JQuery library. The slider and drop-down menu were realized with procedural JavaScript, reading data from the XML file, as well as the implementation of the JQuery library lightbox for the realization of the gallery. Logging in is done with checking using regular expressions in JavaScript.
Technologies – HTML5, CSS3, JavaScript, JQuery, XML, Ajax